Q&A ChatBot/requirements.txt
langchain
langchain-community
langchain-text-splitters
lxml
langchain-openai
python-dotenv
chromadb
langchain-ollama
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq 
